Add -s --single-thread option
To aid in discovering what is going on, the following snippet is often
useful:

```python
import code; code.interact(local=dict(globals(), **locals()))
```

This does however not work if the process runs in a thread pool. This PR
adds the possibility to run the process in a single thread, so
interactive debugging works. The command line options `-s` or
`--single-thread` are now recognized.