Fail on duplicate entries
We had an occurrance where `elliott` was failing to parse
`releases.yml`, as a duplicate key was in there. The validator accepted
this. This commit swaps out the built-in `yaml` module for
`ruamel.yaml`, as is used by our tooling. This will ensure the same
errors are encountered while parsing yaml.