distgit.py: https check requires cert magic, use http
the https check doesn't work in the actual environment it runs (inside
a container) because the distgit server must be validated with an
internal CA, which is tricky to get from outside the firewall for the
container build. we don't need a secure connection just to check for
network reachability so let's not set that up until we need it.